+++ N64.h
@@ -24,19 +24,10 @@ THE SOFTWARE.
 // include guard
 #pragma once
 
-// software version
-#define NINTENDO_VERSION 120
-
 #include <Arduino.h>
 
 //================================================================================
-// Nintendo
+// N64
 //================================================================================
 
-// include all library parts
-#include "Gamecube.h"
-#include "N64.h"
-
-#if (F_CPU != 16000000)
-#error This library only supports 16MHz AVRs
-#endif
+#include "Gamecube_N64.h"
